Extract ModelType interface to a separate file
