+++ api/src/main/kotlin/org/jetbrains/kotlinx/dl/api/inference/keras/loaders/TFModels.kt
@@ -11,12 +11,9 @@ import org.jetbrains.kotlinx.dl.api.core.Sequential
 import org.jetbrains.kotlinx.dl.api.core.loss.Losses
 import org.jetbrains.kotlinx.dl.api.core.metric.Metrics
 import org.jetbrains.kotlinx.dl.api.core.optimizer.Adam
-import org.jetbrains.kotlinx.dl.api.inference.InferenceModel
 import org.jetbrains.kotlinx.dl.api.inference.imagerecognition.ImageRecognitionModel
 import org.jetbrains.kotlinx.dl.api.inference.keras.loadWeights
 import org.jetbrains.kotlinx.dl.dataset.image.ColorMode
-import org.jetbrains.kotlinx.dl.dataset.preprocessor.Preprocessing
-import java.io.File
 
 /**
  * Supported models for inference and transfer learning, trained on ImageNet dataset.
@@ -556,61 +553,5 @@ public object TFModels {
         model.loadWeights(hdfFile)
 
         return ImageRecognitionModel(model, modelType)
-    }
-}
-
-/**
- * Basic interface for models loaded from S3.
- * @param T the type of the basic model for common functionality.
- * @param U the type of the pre-trained model for usage in Easy API.
- */
-public interface ModelType<T : InferenceModel, U : InferenceModel> {
-    /** Relative path to model for local and S3 buckets storages. */
-    public val modelRelativePath: String
-
-    /**
-     * If true it means that the second dimension is related to number of channels in image has short notation as `NCWH`,
-     * otherwise, channels are at the last position and has a short notation as `NHWC`.
-     */
-    public val channelsFirst: Boolean
-
-    /**
-      * An expected channels order for the input image.
-      * Note: the wrong choice of this parameter can significantly impact the model's performance.
-     */
-    public val inputColorMode: ColorMode
-
-    /**
-     * Common preprocessing function for the Neural Networks trained on ImageNet and whose weights are available with the keras.application.
-     *
-     * It takes [data] as input with shape [tensorShape] and applied the specific preprocessing according chosen modelType.
-     *
-     * @param [tensorShape] Should be 3 dimensional array (HWC or CHW format)
-     */
-    public fun preprocessInput(data: FloatArray, tensorShape: LongArray): FloatArray
-
-    /**
-     * Common preprocessing function for the Neural Networks trained on ImageNet and whose weights are available with the keras.application.
-     *
-     * It takes preprocessing pipeline, invoke it and applied the specific preprocessing to the given data.
-     */
-    public fun preprocessInput(imageFile: File, preprocessing: Preprocessing): FloatArray {
-        val (data, shape) = preprocessing(imageFile)
-        return preprocessInput(
-            data,
-            longArrayOf(shape.width!!, shape.height!!, shape.channels!!)
-        )
-    }
-
-    /** Returns the specially prepared pre-trained model of the type U. */
-    public fun pretrainedModel(modelHub: ModelHub): U
-
-    /** Loads the model, identified by this name, from the [modelHub]. */
-    public fun model(modelHub: ModelHub): T {
-        return modelHub.loadModel(this)
-    }
-
-    public fun preInit(): InferenceModel {
-        TODO()
     }
 }
